requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
httpx==0.26.0
python-dotenv==1.0.0
apscheduler==3.10.4
gspread==6.0.0
pydantic-settings==2.1.0
jinja2==3.1.3
aiofiles==23.2.1
python-multipart==0.0.6
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
